Remove the concept of distribution
* We use only one type of distribution thats provided by the dynomite
  manager to keep things simple. Cleanup code related to that
* Reduce some warnings
