+++ src/Makefile.am
@@ -32,11 +32,14 @@ SUBDIRS = hashkit proto event seedsprovider tools entropy
 sbin_PROGRAMS = dynomite dynomite-test
 
 dynomite_SOURCES =			                          \
+        dyn_array.c dyn_array.h		                          \
+        dyn_asciilogo.h                                           \
         dyn_cbuf.h                                                \
-        dyn_crypto.c dyn_crypto.h                                 \
-        dyn_core.c dyn_core.h                                     \
-        dyn_connection.c dyn_connection.h                         \
         dyn_client.c dyn_client.h                                 \
+        dyn_conf.c dyn_conf.h		                          \
+        dyn_connection.c dyn_connection.h                         \
+        dyn_core.c dyn_core.h                                     \
+        dyn_crypto.c dyn_crypto.h                                 \
         dyn_dict_msg_id.h dyn_dict_msg_id.c                       \
         dyn_dnode_client.h dyn_dnode_client.c                     \
         dyn_dnode_msg.c dyn_dnode_msg.h                           \
@@ -44,28 +47,26 @@ dynomite_SOURCES =			                          \
         dyn_dnode_request.c                                       \
         dyn_dnode_proxy.c dyn_dnode_proxy.h                       \
         dyn_histogram.c dyn_histogram.h                           \
-        dyn_server.c dyn_server.h		                  \
         dyn_proxy.c dyn_proxy.h		                          \
         dyn_message.c dyn_message.h	                          \
         dyn_request.c dyn_response_mgr.c	                  \
         dyn_response.c			                          \
         dyn_ring_queue.h dyn_ring_queue.c                         \
         dyn_mbuf.c dyn_mbuf.h		                          \
-        dyn_conf.c dyn_conf.h		                          \
         dyn_node_snitch.c dyn_node_snitch.h                       \
-        dyn_setting.c dyn_setting.h                               \
-        dyn_stats.c dyn_stats.h		                          \
-        dyn_signal.c dyn_signal.h		                  \
-        dyn_token.c dyn_token.h                                   \
         dyn_rbtree.c dyn_rbtree.h		                  \
-        dyn_log.c dyn_log.h		                          \
+        dyn_server.c dyn_server.h		                  \
+        dyn_setting.c dyn_setting.h                               \
+        dyn_signal.c dyn_signal.h		                  \
+        dyn_stats.c dyn_stats.h		                          \
         dyn_string.c dyn_string.h		                  \
-        dyn_array.c dyn_array.h		                          \
+        dyn_token.c dyn_token.h                                   \
+        dyn_log.c dyn_log.h		                          \
         dyn_util.c dyn_util.h		                          \
+        dyn_vnode.c dyn_vnode.h                                   \
         dyn_queue.h			                          \
         dyn_gossip.c dyn_gossip.h                                 \
         dyn_dict.c dyn_dict.h                                     \
-        dyn_asciilogo.h                                           \
         dynomite.c 
 
 dynomite_LDADD = $(top_builddir)/src/hashkit/libhashkit.a
@@ -93,7 +94,7 @@ dynomite_test_SOURCES =                                           \
         dyn_server.c dyn_server.h                                 \
         dyn_proxy.c dyn_proxy.h                                   \
         dyn_message.c dyn_message.h                               \
-        dyn_request.c dyn_response_mgr.c	                  \
+        dyn_request.c dyn_response_mgr.c                          \
         dyn_response.c                                            \
         dyn_ring_queue.h dyn_ring_queue.c                         \
         dyn_mbuf.c dyn_mbuf.h                                     \
@@ -109,6 +110,7 @@ dynomite_test_SOURCES =                                           \
         dyn_array.c dyn_array.h                                   \
         dyn_util.c dyn_util.h                                     \
         dyn_queue.h                                               \
+        dyn_vnode.c dyn_vnode.h                                   \
         dyn_gossip.c dyn_gossip.h                                 \
         dyn_dict.c dyn_dict.h                                     \
         dyn_asciilogo.h                                           \
